requirements.txt
Django==4.2.5
psycopg2-binary==2.9.7
djangorestframework==3.14.0
python-decouple==3.8
requests==2.31.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7